requirements.txt
numpy
pillow
pandas
librosa
sphinx